Maze::Capabilities module dissolved
