+++ lib/maze/browsers.yml
